pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "jellyfish>=1.1.3",
    "matplotlib>=3.10.0",
    "openai>=1.61.0",
    "pandas>=2.2.3",
    "python-levenshtein>=0.26.1",
    "streamlit>=1.41.1",
    "trafilatura>=2.0.0",
    "twilio>=9.4.4",
]
